refactor(match): move Match to conversion/ + rebuild on lattice Prism
Moves `io.github.eschizoid.telescope.Match` → `io.github.eschizoid.telescope.
conversion.Match`. The dispatcher is a paradigm-hop converter (sealed
parent → result), not a navigation primitive — it belongs alongside
`Mapper` / `ForwardMapper` in the conversion sub-package, not at the top
level next to `Telescope` and `Edit`.

Internal rewrite: each `.when(permitClass, handler)` call now creates an
internal `Entry<P, S extends P, R>` holding a lattice `Prism<P, S>` (built
via `Prism.downcast(permitClass)` — the lattice's sealed-narrowing
primitive) plus the per-permit handler. The `Function<P, R>` that
`.exhaustive()` / `.partial()` returns walks the entries and routes each
input through the prism's `getOption(...)` — the first hit wins. The
class-keyed `Map<Class, Function>` table from the original implementation
was a hand-rolled prism dispatch in disguise; this rebuild routes through
the lattice substrate explicitly per the project's lattice-first mantra
("if it could have been a Prism, it should be").

Public API is unchanged — `.of(...) / .when(...) / .exhaustive() /
.partial()` keep their signatures and semantics. Internal types
(`Prism`, `Entry`) stay package-private to the conversion package so the
optic lattice still doesn't leak into the public surface.

Tests unchanged in content; only the import line updates to the new
package location. All 6 `MatchTest` cases pass on the Prism-based
internals (happy path, missing-permit diagnostics, multi-missing
diagnostics, non-sealed root rejection, duplicate-handler rejection,
partial escape hatch).
